optimisations working, project in working state

+++ src/draw_line.c
@@ -6,7 +6,7 @@
 /*   By: crabin <crabin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/27 13:00:46 by crabin            #+#    #+#             */
-/*   Updated: 2025/09/18 20:14:16 by crabin           ###   ########.fr       */
+/*   Updated: 2025/09/19 16:49:14 by crabin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ void init_line(t_line *line, t_node p0, t_node p1)
 void draw_pixel(mlx_image_t	*g_img, int32_t pix_x, int32_t pix_y, uint32_t colour)
 {
 	if (pix_x >= 0 + BORDER && pix_x <= WIDTH - BORDER && pix_y >= 0 + BORDER && pix_y <= HEIGHT - BORDER)
-		mlx_put_pixel(g_img, pix_x, pix_y, colour);
+		mlx_put_pixel(g_img, (uint32_t)pix_x, (uint32_t)pix_y, colour);
 
 }
 
@@ -48,7 +48,7 @@ void    draw_line(t_node p_start, t_node p_end, mlx_image_t	*g_img)
 	p.pix_x = p_start.pix_x;
 	p.pix_y = p_start.pix_y;
 	init_line(&line, p_start, p_end);
-	while (p.pix_x != p_end.pix_x || p.pix_y != p_end.pix_y)
+	while ((int32_t)p.pix_x != (int32_t)p_end.pix_x || (int32_t)p.pix_y != (int32_t)p_end.pix_y)
     {
         draw_pixel(g_img, p.pix_x, p.pix_y, get_colour(p_start, p_end, get_weight(p_start, p, p_end)));
 		e2 = 2 * line.err;
